% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{as_bib}
\alias{as_bib}
\title{Create a bibliography}
\usage{
as_bib(ref, ...)
}
\arguments{
\item{ref}{Object of class `ris`}

\item{...}{Arguments passed from or to methods}
}
\value{
Named list of class `bib` containing formatted bibliography
entries. List names are IDs consisting of first author and year.
}
\description{
Converts an object of class `ris` to a bibliography
}
\references{
American Psychological Association (2019). Publication Manual
of the American Psychological Association (7th Edition). American
Psychological Association.
}
