% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-package.R
\docType{package}
\name{greta-package}
\alias{greta}
\alias{greta-package}
\title{greta: GIS tool of the GRETA project}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org}

}
\keyword{internal}
