% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R, R/app_ui.R
\name{app_server}
\alias{app_server}
\alias{app_ui}
\title{GRETA Shiny base module}
\usage{
app_server(input, output, session)

app_ui(
  theme = NULL,
  preloader = NULL,
  options = NULL,
  sidebar_fixed = FALSE,
  header_fixed = FALSE,
  minified = TRUE,
  controlbar = NULL,
  footer = NULL
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{theme}{Fresh theme to apply to the UI}

\item{preloader}{Preloader to run when the UI is loading}

\item{options}{Further AdminLTE options, see \code{\link[bs4Dash]{dashboardPage}}}

\item{sidebar_fixed}{Whether to fix the sidebar to the screen}

\item{header_fixed}{Whether to fix the header to the screen}

\item{controlbar}{UI to add to the controlbar, see \code{\link[bs4Dash]{dashboardControlbar}}}

\item{footer}{UI to add to the footer, see \code{\link[bs4Dash]{dashboardFooter}}}

\item{minifed}{Whether to minify the sidebar when collapsed}
}
\description{
Create the UI and server function of the GRETA Shiny app.
}
\keyword{internal}
