% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{read_ris}
\alias{read_ris}
\title{Read RIS files}
\usage{
read_ris(file, encoding = "UTF-8")
}
\arguments{
\item{file}{Path to an RIS file}

\item{encoding}{Encoding to be assumed for input strings}
}
\value{
A named list containing reference metadata of the RIS file
}
\description{
Takes a RIS file and parses it as a named list
}
