% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run GRETA Analytics}
\usage{
run_app(
  log = NULL,
  reactlog = FALSE,
  set_python = NULL,
  console = FALSE,
  options = list(),
  ...
)

run_greta(...)
}
\arguments{
\item{log}{Where to output app logs. Accepts a directory path where a log
file is created. If \code{""}, prints directly to the console. If
\code{FALSE}, disables logging. Also accepts a character vector containing
module IDs. In this case, logging is enabled only for these modules.}

\item{reactlog}{Whether to enable logging using the reactlog package.
For debugging purposes.}

\item{set_python}{Python path to set before starting the app. This should
preferably be done before running the function.}

\item{console}{Whether to enable the in-app console. The console can be
accessed by pressing Ctrl + Shift + D and is located in the main module.
For debugging only!}

\item{options}{Options for \code{\link[shiny]{shinyApp}}. Included for use
in \code{electricShine}.}

\item{...}{Further options to be passed to the app. Can be accessed using
\code{getGretaOption}.}
}
\description{
Run GRETA Analytics
}
\examples{
\dontrun{
library(gretan)
gretan::run_app()
gretan::run_greta()
}

}
