% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils_ui.R
\name{tag_to_text}
\alias{tag_to_text}
\title{Converts a shiny.tag object to unformatted raw text}
\usage{
tag_to_text(x, ...)
}
\arguments{
\item{x}{A shiny.tag, shiny.tag.list or list}

\item{...}{Passed to or from other methods}
}
\description{
Converts a shiny.tag object to unformatted raw text
}
