% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-package.R
\docType{package}
\name{greta-package}
\alias{greta}
\alias{greta-package}
\title{greta: The official GIS tool of the GRETA project}
\description{
Visualize and explore the results from the GRETA project. Quickly get insights from GRETA's multi-national survey and case studies. Understand statistical and spatial relationships of energy citizenship in 16 European countries among citizens, businesses and policymakers.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org}

}
\keyword{internal}
