% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run the "greta" Shiny Application}
\usage{
run_app(..., golem_opts = list())

run_greta(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[shiny]{shinyApp}}}

\item{golem_opts}{Options for \code{\link[golem]{with_golem_options}}}
}
\description{
Run the "greta" Shiny Application
}
\examples{
\dontrun{
library(greta)
greta::run_app()
greta::run_greta()
}

}
