% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run GRETA Analytics}
\usage{
run_app(log = NULL, reactlog = FALSE, options = list(), ...)

run_greta(...)
}
\arguments{
\item{log}{Where to output app logs. Accepts a directory path where a log
file is created. If \code{""}, prints directly to the console. If
\code{FALSE}, disables logging. Also accepts a character vector containing
module IDs. In this case, logging is enabled only for these modules.}

\item{reactlog}{Whether to enable logging using the reactlog package.
For debugging purposes.}

\item{options}{Options for \code{\link[shiny]{shinyApp}}. Included for use
in \code{electricShine}.}

\item{...}{Further options to be passed to the app. Can be accessed using
\code{getGretaOption}. Currently this includes the options \code{"console"},
\code{"collect"} and \code{"github_available"}}
}
\description{
Run GRETA Analytics
}
\examples{
\dontrun{
library(gretan)
gretan::run_app()
gretan::run_greta()
}

}
