% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run the "greta" Shiny Application}
\usage{
run_app(..., log = FALSE)

run_greta(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[shiny]{shinyApp}}}

\item{log}{Whether to enable logging using the reactlog package.
For debugging purposes.}
}
\description{
Run the "greta" Shiny Application
}
\examples{
\dontrun{
library(greta)
greta::run_app()
greta::run_greta()
}

}
