% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run GRETA Analytics}
\usage{
run_app(
  log = NULL,
  reactlog = FALSE,
  options = list(),
  prompt = interactive(),
  ...
)

run_greta(
  log = NULL,
  reactlog = FALSE,
  options = list(),
  prompt = interactive(),
  ...
)
}
\arguments{
\item{log}{Where to output app logs. Accepts a directory path where a log
file is created. If \code{""}, prints directly to the console. If
\code{FALSE}, disables logging. Also accepts a character vector containing
module IDs. In this case, logging is enabled only for these modules.}

\item{reactlog}{Whether to enable logging using the reactlog package.
For debugging purposes.}

\item{options}{Options for \code{\link[shiny]{shinyApp}}. Included for use
in \code{electricShine}.}

\item{prompt}{Whether to show a confirmation prompt before downloading
data dependencies.}

\item{...}{Further options to be passed to the app. Can be accessed using
\code{getGretaOption}. This includes the following (mostly internal)
arguments:
\describe{
  \item{\code{console}}{Enables internal live console to execute R
  commands within the app. Strictly for internal use. Console can be
  accessed by pressing Ctrl + Shift + D}
  \item{\code{track}}{Enables user metrics tracking. Requires a valid
  key to upload collected data to Google Drive stored in the environment
  variable \code{GDRIVE_KEY}.} 
}}
}
\description{
Run GRETA Analytics
}
\examples{
\dontrun{
library(gretan)
gretan::run_app()
gretan::run_greta()
}

}
