% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\alias{run_greta}
\title{Run the "greta" Shiny Application}
\usage{
run_app(...)

run_greta(...)
}
\description{
Run the "greta" Shiny Application
}
\examples{
\dontrun{
library(greta)
greta::run_app()
greta::run_greta()
}

}
